requirements.txt
ultralytics
Pillow
opencv-python-headless
numpy
transformers
optimum
optimum-intel[openvino]
openvino
requests
python-multipart
PyYAML
ftfy
git+https://github.com/openai/CLIP.git
onnxslim
onnxruntime
websockets
supervision==0.26.1
trackers==2.0.2rc0
chromadb
ftfy
regex